+++ api/src/graphql/loaders/exif_meta.rs
@@ -1,7 +1,7 @@
 use crate::{
     graphql::loaders::AppLoader,
     graphql::models::ExifMeta as GqlExifMeta,
-    models::exifmeta::{Error as ExifMetaError, ExifMeta},
+    models::exif_meta::{Error as ExifMetaError, ExifMeta},
 };
 use async_graphql::{dataloader::Loader, Result};
 use snafu::prelude::*;
