+++ api/src/models/exif_meta/db.rs
