+++ api/src/models/exif_meta/mod.rs
